# Conflicts:
#	Programming/eco_bike_rental/lib/controller/PaymentController.dart
#	Programming/eco_bike_rental/lib/controller/RentingController.dart
#	Programming/eco_bike_rental/lib/main.dart
#	Programming/eco_bike_rental/lib/view/common/routing.dart
#	Programming/eco_bike_rental/lib/view/payment_view/choose_payment_screen.dart
#	Programming/eco_bike_rental/pubspec.lock

+++ Programming/eco_bike_rental/lib/model/Bike/EBike.dart
@@ -10,7 +10,7 @@ class EBike extends Bike {
   }
 
   EBike.init(int id, String barcode, String color, String category,
-      int bikeValue, int baseRentAmount, int addRentAmount, bool lock)
+      double bikeValue, double baseRentAmount, double addRentAmount, bool lock)
       : super.init(id, barcode, color, category, bikeValue, baseRentAmount,
             addRentAmount, lock);
 }
